fix bug after create comment
